Refactors map-related files and types
